+++ player/src/main/java/fr/unice/polytech/si3/qgl/Mugiwara_Cook/shipmaster/captainNextMove/choice/ChoseDistance.java
@@ -49,6 +49,6 @@ public class ChoseDistance {
     }
 
     public static double distanceBetweenPointAndCheckpoint(Checkpoint checkpoint, double xPoint, double yPoint) {
-        return (double) Math.sqrt(Math.pow(checkpoint.getPosition().getX() - xPoint, 2) + Math.pow(checkpoint.getPosition().getY() - yPoint, 2));
+        return Math.sqrt(Math.pow(checkpoint.getPosition().getX() - xPoint, 2) + Math.pow(checkpoint.getPosition().getY() - yPoint, 2));
     }
 }
